pyproject.toml
[project]
name = "career-intelligence-assistant"
version = "0.1.0"
description = "A RAG-based conversational AI assistant that analyzes resumes against job descriptions"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Web Framework
    "streamlit>=1.40.0",
    
    # LLM & Embeddings
    "openai>=1.58.0",
    "langchain>=0.3.14",
    "langchain-openai>=0.2.14",
    "langchain-community>=0.3.14",
    
    # Vector Store
    "chromadb>=0.5.23",
    "langchain-chroma>=0.1.4",
    
    # Document Processing
    "pypdf>=5.1.0",
    "python-docx>=1.1.2",
    
    # Configuration
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.8.0",
    "mypy>=1.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.13"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
]
